fixed and cleaned up forms on view client page
