+++ client/src/components/clientPage/ViewPage.tsx
@@ -20,6 +20,7 @@ import { useBackendContext } from "../../contexts/hooks/useBackendContext";
 import Comments from "./comments";
 import ChildrenCards from "./childrenCards";
 import Forms from "./forms"
+import { InitialInterview, ExitSurvey, FormItems, SuccessStory } from "./types";
 interface Client {
   age: number;
   attendingSchoolUponEntry: boolean;
@@ -121,42 +122,54 @@ export const ViewPage = () => {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const params = useParams<{ id: string }>();
+  const [formItems, setFormItems] = useState<FormItems[]>([]);
 
   //Toggles visibility
   const [isEditing, setIsEditing] = useState(false);
 
-  //fetches the database for children
-  const fetchChildren = async (id: number) => {
-    try {
-      const response = await backend.get(`/children/${id}`);
-      setChildren(response.data); // Adjust this if the response structure is different
-    } catch (err) {
-      setError(err.message);
-    }
-  };
 
-  //fetches the database for clients
-  const fetchClient = async (id: number) => {
-    try {
-      const response = await backend.get(`/clients/${id}`);
-      setClient(response.data[0]);
-    } catch (err) {
-      setError(err.message);
-    }
-  };
 
   //fetches info from both data
   useEffect(() => {
+      //fetches the database for children
+    const fetchChildren = async (id: number) => {
+      try {
+        const response = await backend.get(`/children/${id}`);
+        setChildren(response.data); // Adjust this if the response structure is different
+      } catch (err) {
+        setError(err.message);
+      }
+    };
+
+    //fetches the database for clients
+    const fetchClient = async (id: number) => {
+      try {
+        const response = await backend.get(`/clients/${id}`);
+        setClient(response.data[0]);
+      } catch (err) {
+        setError(err.message);
+      }
+    };
+
+    const fetchForms = async (id: number) => {
+        try{
+            const response = await backend.get(`/formsCombined/${id}`);
+            setFormItems(response.data);
+        }
+        catch (err) {
+            setError(err.message);
+        }
+    }
     const fetchData = async () => {
       setLoading(true);
       if (params.id) {
         const intId = parseInt(params.id);
-        await Promise.all([fetchChildren(intId), fetchClient(intId)]); // Fetch both routes simultaneously
+        await Promise.all([fetchChildren(intId), fetchClient(intId), fetchForms(intId)]);
       }
       setLoading(false);
     };
     fetchData();
-  }, []); // Fetch data once when the component mounts
+  }, [backend, params.id] );
 
   if (loading) return <Box>Loading...</Box>;
   if (error) return <Box>Error: {error}</Box>;
@@ -205,7 +218,7 @@ export const ViewPage = () => {
             <ChildrenCards items={children} />
           </TabPanel>
           <TabPanel>
-            <Forms/>
+            <Forms forms={[...formItems]}/>
           </TabPanel>
           <TabPanel>
             <Comments clientId={client.id}/>
@@ -583,6 +596,6 @@ export const ViewPage = () => {
         </Box>
       </Box>
     </div>
-    
+
   );
 };
