+++ server/routes/initialInterview.ts
@@ -5,11 +5,21 @@ import { db } from "../db/db-pgp";
 
 export const initialInterviewRouter = Router();
 
-// Get all initial interview form objects
 initialInterviewRouter.get("/", async (req, res) => {
   try {
     const data = await db.query(`SELECT * FROM initial_interview;`);
 
+    res.status(200).json(keysToCamel(data));
+  } catch (err) {
+    res.status(500).send(err.message);
+  }
+
+});
+
+initialInterviewRouter.get("/:id", async (req, res) => {
+  const { id } = req.params;
+  try {
+    const data = await db.query(`SELECT * FROM initial_interview WHERE client_id = $1;`, [id]);
     res.status(200).json(keysToCamel(data));
   } catch (err) {
     res.status(500).send(err.message);
