+++ server/src/app.ts
@@ -23,7 +23,8 @@ import { verifyToken } from "./middleware";
 import { clientDataRouter } from "../routes/clientData";
 import { adminRouter } from "../routes/admin";
 import { intakeStatsFormRouter } from "../routes/intakeStatsForm.js";
-import { calculateMonthlyStats } from "../routes/calculateMonthlyStats"; 
+import { calculateMonthlyStats } from "../routes/calculateMonthlyStats";
+import { formsCombinedRouter } from "../routes/formsCombined.js";
 
 dotenv.config();
 
@@ -71,6 +72,7 @@ app.use("/screenerComment", screenerCommentRouter);
 app.use("/initialInterview", initialInterviewRouter);
 app.use("/intakeStatsForm", intakeStatsFormRouter);
 app.use("/calculateMonthlyStats", calculateMonthlyStats);
+app.use("/formsCombined", formsCombinedRouter);
 
 app.listen(SERVER_PORT, () => {
   console.info(`Server listening on ${SERVER_PORT}`);
